refactor: remove basic TS logic to `@volar/vue-typescript`
