+++ packages/typescript-vue-plugin/tsconfig.build.json
@@ -16,7 +16,7 @@
 			"path": "../shared/tsconfig.build.json"
 		},
 		{
-			"path": "../vscode-vue-languageservice/tsconfig.build.json"
+			"path": "../vue-typescript/tsconfig.build.json"
 		},
 	],
 }
