+++ packages/vscode-vue-languageservice/src/index.ts
@@ -2,10 +2,7 @@ export { getSemanticTokenLegend } from './services/semanticTokens';
 export { getTriggerCharacters } from './services/completion';
 export { margeWorkspaceEdits } from './services/rename';
 export { isRefType, isBlacklistNode } from './services/refAutoClose';
-export * from './utils/sourceMaps';
 export * from './documentService';
 export * from './languageService';
-export * from './sourceFile';
 export * from './types';
-export * from './basicRuntime';
-export * from './typescriptRuntime';
+export * from '@volar/vue-typescript';
