+++ packages/vscode-vue-languageservice/src/services/selectionRanges.ts
@@ -3,7 +3,7 @@ import { transformSelectionRanges } from '@volar/transforms';
 import type * as vscode from 'vscode-languageserver-protocol';
 import { TextDocument } from 'vscode-languageserver-textdocument';
 import type { LanguageServiceHost } from 'vscode-typescript-languageservice';
-import type { SourceFile } from '../sourceFile';
+import type { SourceFile } from '@volar/vue-typescript';
 import type { DocumentServiceRuntimeContext } from '../types';
 import { getDummyTsLs } from '../utils/sharedLs';
 import * as ts2 from 'vscode-typescript-languageservice';
