+++ packages/vue-typescript/README.md
@@ -1,0 +1,3 @@
+# vscode-vue-languageservice
+
+:warning: **This package is planned to be rewritten** :warning:
