+++ packages/vue-typescript/src/index.ts
@@ -1,0 +1,7 @@
+export * from './utils/sourceMaps';
+export * from './utils/string';
+export * from './sourceFile';
+export * from './sourceFiles';
+export * from './types';
+export * from './basicRuntime';
+export * from './typescriptRuntime';
