+++ packages/vue-typescript/src/use/useSfcEntryForTemplateLs.ts
