+++ packages/vue-typescript/src/use/useSfcScript.ts
