+++ packages/vue-typescript/src/utils/localTypes.ts
